Actualizacion proveddores

+++ SistemaTienda.DTO/DetalleCompraCreacionDTO.cs
@@ -13,9 +13,9 @@ namespace SistemaTienda.DTO
         public int IdCompra { get; set; }
         public string Descripcion { get; set; }
         public int Cantidad { get; set; }
-        public double ValorCompra { get; set; }
+        public decimal ValorCompra { get; set; }
         public int ArticuloId { get; set; }
-        public double ValorTotal { get; set; }
-        public double ImpuestoValor { get; set; }
+        public decimal ValorTotal { get; set; }
+        public decimal ImpuestoValor { get; set; }
     }
 }
